requirements.txt
Django==1.9
django-crispy-forms==1.6.0
django-embed-video==1.1.0
Pillow==3.3.1
psycopg2==2.6.2
pytz==2016.6.1
requests==2.11.1
